+++ src/dune_config/dune_config.mli
@@ -116,8 +116,4 @@ val hash : t -> int
 
 val equal : t -> t -> bool
 
-val for_scheduler :
-     t
-  -> Dune_rpc_private.Where.t option
-  -> Dune_stats.t option
-  -> Dune_engine.Scheduler.Config.t
+val for_scheduler : t -> Dune_stats.t option -> Dune_engine.Scheduler.Config.t