+++ src/dune_rules/lib.ml
@@ -311,7 +311,11 @@ module T = struct
 
   let compare (x : t) (y : t) = Id.compare x.unique_id y.unique_id
 
-  let to_dyn t = Lib_name.to_dyn t.name
+  let to_dyn t =
+    Dyn.record
+      [ ("name", Lib_name.to_dyn t.name)
+      ; ("loc", Loc.to_dyn_hum (Lib_info.loc t.info))
+      ]
 end
 
 include T
@@ -1757,7 +1761,7 @@ module DB = struct
     | None ->
       Code_error.raise "Lib.DB.get_compile_info got library that doesn't exist"
         [ ("name", Lib_name.to_dyn name) ]
-    | Some lib -> Compile.for_lib ~allow_overlaps t lib
+    | Some lib -> (lib, Compile.for_lib ~allow_overlaps t lib)
 
   let resolve_user_written_deps_for_exes t exes ?(allow_overlaps = false)
       ?(forbidden_libraries = []) deps ~pps ~dune_version =