+++ src/dune_rules/ocamldep.ml
@@ -70,12 +70,11 @@ let interpret_deps md ~unit deps =
   | None -> deps
   | Some m -> m :: deps
 
-let deps_of md ~ml_kind unit =
-  let modules = md.modules in
-  let sctx = md.sctx in
+let deps_of
+    ({ sandbox; modules; sctx; dir; obj_dir; vimpl = _; stdlib = _ } as md)
+    ~ml_kind unit =
   let source = Option.value_exn (Module.source unit ~ml_kind) in
-  let dir = md.dir in
-  let dep = Obj_dir.Module.dep md.obj_dir in
+  let dep = Obj_dir.Module.dep obj_dir in
   let context = SC.context sctx in
   let parse_module_names = parse_module_names ~modules in
   let all_deps_file = dep (Transitive (unit, ml_kind)) in
@@ -86,16 +85,16 @@ let deps_of md ~ml_kind unit =
       (let open Action_builder.With_targets.O in
       let flags, sandbox =
         Option.value (Module.pp_flags unit)
-          ~default:(Action_builder.return [], md.sandbox)
+          ~default:(Action_builder.return [], sandbox)
       in
       Command.run context.ocamldep
         ~dir:(Path.build context.build_dir)
+        ~stdout_to:ocamldep_output
         [ A "-modules"
         ; Command.Args.dyn flags
         ; Command.Ml_kind.flag ml_kind
         ; Dep (Module.File.path source)
         ]
-        ~stdout_to:ocamldep_output
       >>| Action.Full.add_sandbox sandbox)
   in
   let build_paths dependencies =
@@ -115,10 +114,10 @@ let deps_of md ~ml_kind unit =
     let open Action_builder.O in
     let paths =
       let+ lines = Action_builder.lines_of (Path.build ocamldep_output) in
-      lines
-      |> parse_deps_exn ~file:(Module.File.path source)
-      |> interpret_deps md ~unit
-      |> fun modules ->
+      let modules =
+        parse_deps_exn ~file:(Module.File.path source) lines
+        |> interpret_deps md ~unit
+      in
       ( build_paths modules
       , List.map modules ~f:(fun m -> Module_name.to_string (Module.name m)) )
     in