+++ src/dune_rules/odoc.mli
@@ -1,9 +1,9 @@
 (** Odoc rules *)
 
 open Import
-open Dune_file
 
-val setup_library_odoc_rules : Compilation_context.t -> Library.t -> unit Memo.t
+val setup_library_odoc_rules :
+  Compilation_context.t -> Lib.Local.t -> unit Memo.t
 
 val gen_project_rules : Super_context.t -> Dune_project.t -> unit Memo.t