+++ src/dune_rules/toplevel.ml
@@ -112,7 +112,7 @@ let setup_rules_and_return_exe_path t =
   let open Memo.O in
   let linkage = Exe.Linkage.custom (Compilation_context.context t.cctx) in
   let program = Source.program t.source in
-  let* () =
+  let* (_ : Exe.dep_graphs) =
     Exe.build_and_link t.cctx ~program ~linkages:[ linkage ]
       ~link_args:
         (Action_builder.return
@@ -136,7 +136,7 @@ module Stanza = struct
     let open Memo.O in
     let source = Source.of_stanza ~dir ~toplevel in
     let* expander = Super_context.expander sctx ~dir in
-    let scope = Super_context.find_scope_by_dir sctx dir in
+    let* scope = Scope.DB.find_by_dir dir in
     let dune_version = Scope.project scope |> Dune_project.dune_version in
     let pps =
       match toplevel.pps with